+++ src/interfaces/ActionItemPropsInterface.tsx
@@ -1,0 +1,8 @@
+interface ActionItemProps {
+    name: string;
+    image: string;
+    level: string;
+    onAdd: () => void;
+}
+
+export default ActionItemProps;
